[MISP] Adds propagate labels from Event to Attribute option #3495
Currently the Attributes tags take precedence over the Events tags (if attribute has tags)

This commit allows for the option to allow the labels that are present on an Event (OpenCTI Reports) to apply on top of the labels set by the Attributes (Observables and/or Indicators).
While not changing behaviour by introducting a setting that defaults to the previous behaviour.